requirements.txt
python-telegram-bot==13.12
validators==0.20.0
sentry-sdk==1.5.12
requests==2.28.0
Pillow==9.1.1
